The AI is now functional.
I created the entierty of the AI State scripts and the guards are
patroling as they should.

I need to fiddle with the details inorder to make it look better.

The inishal paramiters that i tryed for the guard sight may need to be
changed because they curently do not detect the player.